Feat: Cambios en los modelos

+++ PracticaGestionInventario/Models/EntradasDetalles.cs
@@ -8,7 +8,7 @@ namespace PracticaGestionInventario.Models
         public int Id { get; set; }
         public int EntradaId { get; set; }
         public int ProductoId { get; set; }
-        public decimal Cantidad { get; set; }
-        public decimal Costo { get; set; }
+        public int Cantidad { get; set; }
+        public double Costo { get; set; }
     }
 }
